+++ src/engine/session/commands.ts
@@ -14,8 +14,8 @@ const addSession = (s: Session) => {
   pubkey.set(s.pubkey)
 }
 
-export const loginWithPrivateKey = privkey =>
-  addSession({method: "privkey", pubkey: getPublicKey(privkey), privkey})
+export const loginWithPrivateKey = (privkey, ...extra) =>
+  addSession({method: "privkey", pubkey: getPublicKey(privkey), privkey, ...extra})
 
 export const loginWithPublicKey = pubkey => addSession({method: "pubkey", pubkey})